Fix definition of vnode's isstable (and move out of chord.h); this is for stabilizing the predecessor.
Sometimes we fix our successor as part of stabilizing the finger table
(perhaps an implementation artifact) --- notify new successor if we
change it in updatefinger. This was causing us not to restabilize
after node failure. May want to revisit this later and not maintain
successor as part of the finger_table; instead perhaps a separate class
to track successor and predecessor.

Simplify stabilize_finger. Now only two things to worry about. If the
finger is dead, quickly find a temporary replacement and then issue an
RPC to find the true new successor. If the finger is alive, check its
correctness by asking it for its predecessor. Also, skip over fingers
that are the same when stabilizing. This makes things faster again.
Add some comments to help make it clear what's going on.