added code and interface that optionally allow dhash to directly go to what it believes is the successor of a block ID on fetch or insert, instead of sending a RPC to the predecessor first. dhashclient interface is modified so caller can specify if dhash should do a full lookup first or go directly to the successor.
this does not break correctness because on fetch, if the block is not there,
caller can try again with full lookup. on insert, if the node is not
responsible for storing the block, it will answer the RPC with a RETRY reply.

this change helps applications (e.g. ivy) that insert and fetch public
key blocks very often. in this case the node responsible for the public
key block is likely to be in a dhash node's location cache.