*INTERFACE CHANGE*
dhashclient.insert now returns a insert_info structure instead of just
a key. The insert_info contains the key along with the node where the
data item ended up. I fixed anything that was in the repository, but if your
software uses dhashclient and isn't in sfsnet you'll have to update.