update test_merkle_syncer to new merkle interface
bug fix in merkle_syncer: between getblockrange and getblockrange_cb
the remote side was sending all the blocks over and thus causing the
synchronization to be finished.  Then when the getblockrange_cb came
it pointed at freed memory.

more cleanly separate the merkle code and dhash code by way of the
addHandler callback.