partial fix to append blocks
reduce the size of many RPCs by leaving out the chordID.
chord_node_wire is a new type without a chordID. use it in RPCs.
chord_node is the same as before, but is used internally for locations, etc.
we can regenerate the chordID for chord_node when necessary using
hostname/port/vnode.