Factor of 3 speedup on Jinyang's test case (from 102r 19u 91s to 28r 26u 2s on tears).
I inserted dummies for a lot of functions that aren't
necessary in single-process programs.  They're clearly marked.

Thread communication was using a hash table of size 13 to
hold the threads waiting to talk to other threads.  Now the
hash table is size 1009, which should make the communications
a bit closer to constant time.

Most embarrassing of all, though, is that thread teardown was
an O(number-of-threads) operation.  That explains Frans's
12-hour 1000-node simulation.  Now it's O(1), which should
work a lot better.

You need a new libthread.a to get the effects of the
various speedups.  Frenulum, tears, and amsterdam have them
installed, and I supfixnscan'ed on frenulum, so sup
should be enough to get your machine updated.
(If you use an old libthread.a, it'll still work, you
just won't see a speedup.)

Linux guys can copy the new sources and Makefile out
of ~rsc/libthread.

Now there aren't any obvious hot spots in the trace.
These are the top ten:

 2.6%	/usr/home/am3/rsc/p2psim/p2psim:Packet::~Packet(void)
 2.6%	/usr/home/am3/rsc/p2psim/p2psim:_threadready
 2.8%	/usr/home/am3/rsc/p2psim/p2psim:EventQueue::advance(void)
 3.3%	/usr/home/am3/rsc/p2psim/p2psim:altexec
 3.4%	/usr/home/am3/rsc/p2psim/p2psim:LocTable::findsuccessor(unsigned long long)
 4.0%	/usr/home/am3/rsc/p2psim/p2psim:_schedinit
 5.4%	/usr/home/am3/rsc/p2psim/p2psim:alt
 5.4%	/usr/home/am3/rsc/p2psim/p2psim:Euclidean::latency(Node *, Node *)
 6.5%	/usr/home/am3/rsc/p2psim/p2psim:_init
 6.5%	/usr/home/am3/rsc/p2psim/p2psim:_sched

I feel a little bad about alt and _sched being as big as they
are, but it's not clear how to make them smaller -- the
simulator just does a hell of a lot of communication
and scheduling.