Add chord-monitor that listens for incoming TCP connections from lsd's to receive copies of each vnode's state, dump the state to a file, and then periodically run stable.pl to determine if the ring is stable.  Make sure this is running before any lsd's are started.
Add a '-w' flag to lsd that tells it the host:port that chord-monitor
is listening on and causes it to dump the "locationtable" to that
host:port.

Update the print method in chord/vnode/fingerlike classes to take a strbuf,
so that USR1 and -w can both use the same print method.