Support for asynchronous RPCs.  Robert, please take a look.
doRPC has not changed.

asyncRPC is new.  It gets the same parameters as doRPC but it returns
immediately a pointer to a RPCHandle which contains (among other things) the
pointers to args and ret that were originally passed to asyncRPC.

select() takes an RPCSet* (which is a set<RPCHandle*>) as its only parameter
and returns one RPCHandle*.  select() will remove the returned RPCHandle* from
the set, but it will NOT delete the returned RPCHandle itself.  This is because
the application now needs to access the RPCHandle's args and ret members to
figure out which RPC returned, and then deal with it.  It is up to the
application to properly delete RPCHandle objects.  Failing to do so will result
in severe punishment.