Changed the asyncRPC plan.
asyncRPC no longer returns an RPCHandle, but, rather, an unsigned---the
``token''.  This token is used just like the RPCHandle was used before.  For
example:

  RPCSet *rpcset = new RPCSet;
  unsigned token = asyncRPC(wkn, &Kademlia::do_lookup, &la, &lr);
  rpcset->insert(token);
  unsigned ready = select(rpcset);
  assert(ready == token);

select() returns one token that is ready and takes it out of the set.

It is up to the application to keep track of these tokens and to map them to
any data is needed when the RPC completes.

The application can choose its own token by passing it as an (optional) last
parameter to asyncRPC().

  unsigned token = asyncRPC(wkn, &Kademlia::do_lookup, &la, &lr, 18);
  // token == 18

Thomer's wrath will descend upon you when you pick a token that is associated
with an RPC that's still out there.

You can cancel an RPC by calling cancelRPC(token).  This destroys all data
associated with the RPC.