okay, this is the changes to the keyhash block object we discussed earlier with michael walfish. keyhash block is now a object stored under the hash (publickey, salt), where salt is 20 bytes.
the keyhash object's structure is defined in dhashclient.h, as
keyhash_payload class.  all the related methods are defined in
there.  for example, methods to sign, verify, compute chord id,
encode, decode, etc.

for example, the call

  keyhash_payload ().id (pk)

where pk is a sfs_pubkey2, returns the ID of a keyhash block with
out a salt. where

  keyhash_payload (salt).id (pk)

would return the ID of a keyhash block with salt.  there are other
constructors of course that actually take in a version and a data
buffer.

tested on ron.