New story for system-wide parameters to a protocol.
p2psim now requires three files.  The third file is a file with per-protocol
parameters.  Take a look at protocol.txt for an example.

Kademlia is a working example.  ProtocolFactory passes an Args object to the
constructor.  The constructor can then get values out.

Jinyang now needs to take the global variables out of main.C and modify
ProtocolFactory to pass the Args object to the constructor of all Chord-related
classes.