change dbwrite () in dhash/server.C in favor of db_write_immutable ().
fixed a memory leak in lsd/gateway... was creating a gateway object,
and an axprt_stream object (w/ 1Mb of buffer), each time a connection
to lsd was made via dhashclient. this is the good news.

the bad news is that we still have memory leaks. 500M taken by lsd
after my experiment finishes...