Hello, my name is Lucifer.  Welcome to GCC-Hell.  Let me give you a quick tour around the place.
On your left you see GCC2.  In GCC2, hash_map is included with
#include <hash_map> and lives in namespace std.

On your right you see GCC3.  In GCC3, hash_map is included with
#include <ext/hash_map> and lives in namespace __gnu_cxx.

Oh, and that little fellow in the middle?  Ignore him.  He's been running back
and forth for the past few hours mumbling shit about configure.in.