in chord/server.C, if notify fails, remove the node we tried to send the notify request to. this allows a lsd behind a NAT to remove location objects outside of the NAT from its table, eventually causing it to function as a ring of size 1.
emil, could you check if this is okay?

changed gateway.C so proxy mode occurs when the ring is size 1, and a
proxy server was specified.

this set of changes allow a lsd ring to be able to transition between
fully connected, disconnected, and behind a proxy easily, w/o having to
be killed and restarted.

we still have problems with
  - address/interface changes

  - moving data written to local lsd while in disconnected mode to the
    rest of the network when we hook up to a proxy server. but i think
    it's going to turn out that this can be done more effectively by
    the application, instead of dhash... we will see.