FINALLY figured out what's going on with exiting.  Threaded() should not call threadexits(0) as this interferes with clean destructor behavior.
Cleaned up Jinyang's EventQueue hack.

Plugged various memory leaks with one notable exception: As things are now,
there's a small leak on every single RPC.  I don't (yet) know how to solve
this, as this is part of the redesign that was done to accomodate for asyncRPC.

Fixed documentation in sc.pl.