Some changes to help scalability. sim isn't perfect, but it's a bit better now.
changes include:
o The number of successors is now configurable on the commandline.

o There is now an option to "fake" document list operations. setting this
  to true makes a difference for large simulations (e.g., 10^6 documents).

o The size of the location cache and the number of fingers are configurable
  on the commandline.

o The chief thing that makes this run faster is that nodes no
  longer poll their request queues unless there's a request actually in the
  queue.

o Document lists are now ordered; that also cuts down
  on some operations as well as on duplicates in the doc list.

o The node hashtable is smaller; makes it faster to iterate through the
  table.

o Fingers and successors are now pinned into the location table; eviction
  happens LRU among non-fingers and non-successors.

o If a node N communicates with another node M, N adds M to its location
  table, always (previously, sim missed some cases).

o If a node N tries to communicate with a down node M, then N evicts M from
  its location table after a timeout. Previously, sim missed some cases.

o Fixed a few bugs, like failing to free memory and failing to initialize
  a variable or two.

o Haven't touched the traffic generator; one might want to change the
  document and node lists to, e.g., a hash table instead of an unordered
  list. O(log n) would be an improvement over O(n^2). Currently, I'm using
  something in Perl....  Event generation with trafficgen will be a
  bottleneck, especially since the maximum number of documents is limited.