Add failure models.  When an RPC was sent to a dead node, the returning packet is delayed according to a failure model.  The specific failure model is defined with keyword ``failure_model'' in the topology file and applies to all nodes. Upon request, we could make a failure model local to a node.
A FailureModel needs to implement failure_latency.  It can inspect the packet
and make any decision it wants on how long to delay the packet.  This interface
is very similar to Topology.

Only one failure model is now implemented: the NullFailureModel, which adds
zero latency.

Failure models add overhead because failed packets are sent back to the network
layer and are rescheduled through NetEvent and the EventQueue. To avoid this
overhead, p2psim can be run with -f to ignore failure models altogether.

Adjusted sc.pl.  You need to run it again or you will be punished.