Fix the initstate story.  I dislike the fact that observers do inistate.
Protocol now has a virtual function initstate.  After everything has been set
up, but before the eventgenerator are put to work, initstate(set<Protocol*>*) is called on
all Protocols in the system.

Note that the Protocol needs to use proto_name() to figure out whether using
a particular instance of a Protocol is appropriate.

Take a look at Kademlia::initstate() for an example.

getallips() and getallprotocols() now return a set, not a list.