Various changes.
Rather than shitting around with splitting/collapsing k-buckets tree, just
generate the entire tree at constructor time and that's it.  That also solves
the paradox in the Kademlia paper.  (On the hand it says to put nodes in the
replacement cache when the k-bucket is full, but at the same we're supposed to
split the k-bucket when it's full.  Uh?)

Let find_node use its own recursive tree traversal thingy.

Impose a limit (k) on the size of replacement cache.

The result of all this is that things are much faster and easier to read.