ChurnFileEventGenerator = cross between the churn and file event generators. Allows you to use a trace to drive the churns, while still using random lookups file format:
<ip> {dead/alive} <time1> <time2> . . . <timeN>

the dead or alive signals whether or not a node is alive at the beginning
of the test.  The times signal when a node changes state -- that is, if a node
starts out as alive, then time1 is the time when it first dies.

Note: if you want to use this in your protocol, and you use an initstate
function, you'll have to make sure it only initializes nodes that are alive at
the beginning of the test (dead nodes will fail a node()->alive() check).

Besides the filename ("file"), it takes the same params as the
churneventgenerator (it's a subclass), but ignores the lifemean and deathmean
params.