I think I fixed the join bug. It turned up only in the recursive case because the iterative code calls insert on each intermediate hop. So even if the lookup comes back to us we have a "successor" and can call it done.
In the recursive case, we don't have nodes in the bag, so we time out
when the request comes around again.

My solution is to look up myID - 1 so that we never get consulted and
to modify find_route to return successors. I just stick the whole
route and the succs inthe bag.

This might not work if overshot is turned out. If that turns out to be
a problem we ould just add the well known node before we join. Then
we'll have a succ. (albeit a bad one) just in case.

NOT TESTED ON PLANETLAB