Hack run-simulations to be able to iterate through or randomly pick from events parameters, in addition to protocol parameters.  It knows about lifemean, deathmean, lookupmean, exittime, and stattime.  Just put values in your "args" file like you would any other of your protocol parameters.  For example, if you want to explore a churn mean of 1 hour vs. 100 seconds, with a lookup mean of 10 minutes, and have the test always last 12 lifetimes (collecting stats for half of that time), you would put the following in your args file:
lifemean 3600000 100000
deathmean =lifemean 1
lookupmean 600000
exittime *lifemean 12
stattime *exittime .5

If values for these parameters are given via the command line, those
command line values override the values in the args file.

By convention, it would be best to put these events parameters in the above
order at the top of your args file.  That will be easiest for graphing later
on.