Slow down stabilization from break-neck speed to something approaching reasonable.
We target a successor stabilization every 20 seconds, and a full
set of fingers stabilized about once a minute.  Set a faster initial
rate so that new nodes are up to date pretty quickly.  Will backoff
to the target times, and slower if necessary.

The current numbers could probably be tuned more.