Less verbose in the safe case.  Avoid a null pointer deref in the process. Clean up the output sent to the remote end in the case of missing articles.
Right now, missing articles are only deleted from the header database,
and not from the individual group article lists.  This is because the
article list is one big string which is parsed with rxx's.  Urgh.  But
if the article is not in the header database, don't generate anything
at all about the article and don't crash.