Split dhash into a chord class that just maintains node lists and time, and a dhash class that inherits from chord that provides storage. This is in prep for total recall implementation.
Also, add a cached copy of the block on the actual successor of
the block; this reduces b/w used most of the time during repair,
but increases storage bandwidth.  Does not yet account for additional
disk space required to store this copy... didn't want to think about
the GC issues, if any.