Keep track of total_unavailability.  Adds a bunch of book-keeping code: When nodes fail, decrement their blocks' availability and when a block becomes unavailable (fewer than the necessary number of pieces to reconstruct), start counting its unavailability timer until it becomes available again.
Ensure that in Total Recall's block repair algorithm, you don't try and
re-replicate on a node that is already in the inode --- such nodes already
have the block or are dead.  Interestingly, you might try to re-replicate to
a node that has the block but you just forgot about it.