num_fetch should be at least as many copies as were written, since that governs the number of successors that are retrieved.  In the case of content hash blocks, set that to be the number of successors (retrieved via chord.nsucc in configurator).
Also, dispatch min_fetch () fetch requests in parallel initially, not
num_fetch ().  Perhaps this naming is slightly confusing.

(This should help correctness but we still need to arrange it so that the
returned successor list has an indication of holes or some probability of
whether the block is actually on the node in the successor list for
performance.)