Add a seteofcb front-end in dhashclient that passes through to aclnt.
When an RPC that is larger than the max allowed size (hard-coded in
dhashclient, but we could make it parameterizable), aclnt closes the
connection causing all future RPC requests to fail.  However, because
it calls an eofcb first, we can detect this and re-open the connection
manually.  In UsenetDHT, this was being triggered occasionally by
extremely large posts (apparently to news.admin.net-abuse.bulletins).

This behavior can be tested by having dbm try and store a sufficiently
large block (e.g. 1024*1025 + 1).