Force linker to include dhblock.o with syncd.
Configurator keys are initialized with static initializers in
object files; dhash.replica is set by dhblock.o.  Unfortunately,
there were no calls to any methods in dhblock.o so the initializer
was not getting included.  Add a dummy call to force that to happen.
(Not clear if this was actually causing a problem for NOAUTH blocks
or not.)