Remove ancient resendRPC/rexmit interface.
The old interface had doRPC return a "sequence number" (really just
an identifier) that a caller could later use to ask the RPC subsystem to
explicitly retransmit that particular RPC.  Sometime ago, we replaced
this with the ability to specify a timeout callback: if specified,
the caller would be called back before STP would retransmit the RPC
and return a bool to decide whether to actually retransmit.
Accordion uses this new interface to turn off retransmits.

At this point, the return type of doRPC could be switched back to void.