adbd/dbfe: Enable DB_TXN_WRITE_NOSYNC for performance.
This flag tells BerkeleyDB to allow the transaction to commit
even if the OS has not guaranteed that the data has made it
to stable storage.  Reduces lsd/dhash amortized write operation
cost from 40ms to 4ms on oc-5.  Raw adbd writes "complete"
in approx 0.1ms so most of that overhead is probably memory copies
in lsd.

Other options might to be to log just in memory or to use
DB_TXN_NOSYNC; the documentation suggests this might be the safest
compromise, protecting against application crashes but not OS
crashes.

http://www.sleepycat.com/docs/gsg_txn/C/usingtxns.html#nodurabletxn