Improve startup performance with large content hash databases.
During initial start-up, a merkle tree is constructed for
blocks stored locally, for each datatype.  This requires
iterating over the database (with a BerkeleyDB cursor).
Unfortunately, when the database is large (e.g., millions
of keys with (8K)/7-ish bytes of data), this can be inefficient.
One hypothesis for what is happening is that BerkeleyDB
potentially has to read lots of data off the disk: even though
we tell BerkeleyDB not to return the data, the underlying
BTree stores key+data on the same physical page.  Further,
cursors for a BTree read in order, resulting in lots of seeks
since data is inserted randomly.

keyauxdb class manages a simple file which contains key+auxdata and
is designed simply to be streamed from disk.  This first pass
implementation does not support deletion or updates and is only
used for content hash databases.  This may in fact be sufficient,
as auxdata (e.g., for noauth blocks) is already stored in a
separate database that is much smaller.  Of course, this won't
be seek-free either, since it is a regular file and data blocks
are not guaranteed to be laid out sequentially.  But it is much
smaller and has more keys per disk block.

Empirically, the new implementation may issue about an order of
magnitude fewer read system calls.  It is consistently faster
than the BerkeleyDB cursor but the performance of BDB seems to
vary widely depdning on how much info may be stored in its own
cache and in the buffer cache.

Specific changes:
* Extend adbd to use keyauxdb, "migrating" old db if neceessary.
* Update adb::getkeys API again to allow ordered/unordered calls.
* Change getkeys callbacks to get a generic "continuation" to give
  back to adbd instead of specifying a "start"; track continuations
  with ids in libadb getkeys calls.
* pmaint prefers an ordered iteration; dhblock_*_srv does not.
* Some dead code removal in adbd.C.