Fix ACLNT_TRACE (?!)
It would appear that since the introduction of TRANSPORTPROC, our
hacked ACLNT_TRACE has been printing out the wrong information
when trace level was set higher than 5.  The "res" pointer that
was being printed was a transport dorpc_res, but being interpreted
as a "res" for the real RPC program.  This fix interposes the printing
right before the real cb is called as opposed to before doRPC_cb.