Fix omission of last predecessor in pred_list::fill_nodelist*.
The old code assumed that the local node should be the zero-th
element but did not account for it in sizing the response array.
This meant that in the normal case, one predecessor was always
missing.

Similarly, if there are no predecessors, the local node would not
be returned at all.  Fixing this part makes the previous commit
superfluous but merely safe.

As part of this, make sure sampler's view of the predecessor list
skips the 0th entry.  (This is not really necessary for the
correctness of the sampler code, which would have been ok now that
the last predecessor is not missing.)