Allow an explicit completion callback for merkle syncer.
Merkle syncer provides a polling and callback mechanism for
determining when a sync is done. The old callback was overloaded as
a third parameter passed back in the missing function.  I don't
think that will work with TAME; this patch adds an explicit cbi
which can optionally be provided by the caller when the sync is
started.  An error code will be called back; 0 if nothing went wrong.