Add guards to prevent IDA crashes due to short inputs.
Jeremy has observed over the past year several cases where
fragments have been too short; this would cause IDA's
reconstruction code to make an out-of-bounds memory access
and, if --with-dmalloc, generate an assertion failure in lsd.
This patch adds several guards to prevent the crash and
gracefully fail the reconstruction.

Unfortunately, we don't know why the inputs might be too short.
It'd be nice also if we could find corrupted fragments and remove
them from the system.