Update all DHash constructors to have consistent parameter names.
When we updated to using adbd, we simply co-opted the str, str
parameters passed to DHash constructors from dbname + dbext,
(which specified the path plus type of database) to dbsock + dbname
(which specifies adbd's unix domain socket and the adbd namespace);
we did not update the parameter names to reflect this new usage.
This was confusing so this patch brings the names in-line with
the current usage.