Add ADBPROC_GETSPACEINFO to return full path and hasaux of a db.
adbd now internally uses full path (as returned by getcwd)
instead of just the -d argument passed in on the command line.
GETSPACEINFO is used to get the full path of the directory
used to house data about a given namespace (constructed by
appending the namespace name to the base path), along with
whether the namespace was initialized to use aux data.

The main idea behind this RPC is to allow maintd a clean
way to find out where adbd has hidden the Merkle tree for
a given node/ctype.