Wait for getkeyranges to complete before declaring merkle_syncer done.
Jeremy noted:

    When node A syncs with its neighbor B, and initially has an
    empty merkle_tree for B (i.e., A's root for B's tree is a
    leaf), then merkle_syncer::sendnode_cb does not add it to the
    "st" queue.  This means that while B is sending back its full
    list of keys in the specified range, merkle_syncer::next is
    calling setdone(), which cause the process to be aborted when
    the next successor is chosen during the next synchronization
    interval.

This patch keeps track of how many getkeyranges are in progress,
and waits until they all finish before allowing the overall sync
to be marked as completed.

Move the declaration of compare_nodes into merkle_syncer proper
so that it can interact with and update the count of outstanding
getkeyranges as necessary.

Remove the declaration of compare_keylists from merkle_syncer.h
since it is not needed by anything else.

test_merkle_syncer passes with this patch, though that doesn't
yet guarantee correctness.