Test merkle_nodes for equality before doing any other work.
merkle_syncer did not compare the hashes of nodes unless doing
index-index comparisons.  In that case, it did an extra 64
sendnodes unnecessarily.  In that, and all other cases, it
would read keys off the disk unnecessarily, even if everything
turned out to be the same.  By comparing the hashes first,
we can avoid potentially a lot of network and disk I/O.

Originally noted by Jeremy.