Break memory/fd leak due to refcount cycle in merkle_syncer/maintd.
The use of TAME comes with some risks; the refcounted
tame closure for maint_sync contains a merkle syncer ptr,
the merkle syncer contains a pointer to a completion callback,
the completion callback contains a reference to the closure.
Break the loop by clearing the completion callback pointer
in merkle_syncer after it is called.

Thanks to Max for pointing out the loop.  Thanks to Joel Spolsky
for the Law of Leaky Abstractions.
http://www.joelonsoftware.com/articles/LeakyAbstractions.html