A basic Passing Tone implementation.
Maintain a list of objects in the current range of responsibility
that are missing locally from the local node relative to the
current successor and predecessor.  Keep this list in memory and
return it to lsd in response to MAINTPROC_GETREPAIRS.

Synchronization is handled using Merkle trees, mirroring Carbonite.
Like Carbonite, use the synchronization to update adbd's view of
block placement.

This implementation only starts a new synchronization round is
started only when the repair queue has been emptied.  This means
that changes in successor, predecessor or range of responsibility
will only be updated after existing repairs are handled.  Of
course, at this point, lsd does not ever call MAINTPROC_GETREPAIRS,
so only a single sync round will occur.

Because the list of differences is only in memory, if maintd
crashes, bandwidth will be spent re-learning these differences.
However, writing this to disk seems like more trouble than it is
worth at the moment.