Add a sync method to merkle_tree_disk to take real snapshots.
Use snapshots to ensure that everyone has a consistent view of
the merkle tree.

merkle_tree_disk now has a sync method that either creates a snapshot
(when writable) or reloads a snapshot (when read-only).  For writers,
the sync method: takes out a lock file (to prevent the reader from
opening files during the snapshot process); closes the current .mrk
files; makes copies of the .mrk files to .mrk.ro; re-opens the .mrk
files for read/write.  If a merkle_tree_disk is not opened as a
writer, it will close and re-open its FILE *s, using the .ro files.

Call this sync method in various places: flush the adbd tree to disk
based on the repair timer in DHash, sync the local read-only tree
before starting a maintd repair cycle, whenever a writable merkle_tree
is closed.

Update the test code to call sync and also to clean-up .ro files.