Actually start maintenance for noauth blocks.
Maintenance is started when all dhblock_*_srvs have finished
initializing, which a dhash object notices when it collects all
the completion callbacks from the constructors.  It happens that the
current initialization doesn't do anything async, so the callbacks
were being called immediately: in particular, they were being
called before the newly object was inserted into the list of block
servers.  This had the effect of calling start_maint () before the
last block server to be created was added to the server list.

Fix this by scheduling the callbacks to be called using delaycb
instead of calling them directly.