Prevent unnecessary re-open on merkle_tree_disk destruction.
In addition to being "nicer", it seems that in certain exit conditions
(e.g., unclean fatal exit), the order of destruction of static
resources and heap-allocated objects can be such that the static rxx
used in merkle_tree_disk.C's open_file gets freed/corrupted before
an outstanding merkle_tree_disk is destroyed, resulting in:

    assertion "r->magic == rcbase::magicval" failed at /home/sit/work/sfslite/sfslite2/async/rxx.C:64

Um.  Yeah.  Rather than figure out how to order things correctly (or
if this is really the problem), let's just not rely on that static
resource during the destructor.