Remove fatals/asserts from merkle synchronization.
The code as originally written made the assumption that the
client and server trees would remain consistent during the
synchronization process.  This may have been true in the past
for small trees/fast synchronizations but certainly isn't true
now.  Inconsistencies were causing various checks to fail and
crash maintd.

Therefore, this change ensures that any time something goes wrong, we
simply ignore that portion of the tree and try to carry on.  It is not
necessary to identify all differences on a single pass (or perhaps
even possible, since the trees keep changing)---subsequent
synchronization iterations will hopefully detect the missed
differences.  (If the system quiesces, they definitely will.)