Add suggestion of repair source to repairs generated by maintd.
The current suggestion is the first successor with a copy of the
key.  Carbonite only returns a suggestion if there's an aux key
(implying that the objects are mutable); this is to best support
the potential of fragmentation, which makes the most sense for
immutable content-hash objects. (I'm not ready to think about
with regards to mutable objects).   However, Passing Tone always
includes a hint as to whether to get it from the successor or the
predecessor.