Track repairs by object id + (source +) destination; not just id.
Code previously tried to ensure correctness by only allowing
one outstanding repair per object; but this was only shooting itself
in the foot because noauth servers in fact legitimately attempt
enqueue more than one repair job for the same object (but different
destinations).  So, instead of just tracking the keys, construct
a short description of the repair job as a string and track by that.

Change some logging from trace to info so we see more repair status
at the default log level.  Also, only actually log that we added
a repair when we do, not when we were asked to and then decide not to.