Add DHASHCLIENT_SKIP_LOOKUP option to dhashcli::retrieve.
When used with benjie's ptr<chordID> guess (and DHASH_GUESS_SUPPLIED),
this allows a dhash user to provide a specific node to retrieve from.
This is primarily aimed to be useful for repairs of replicated
(non-fragmented) objects where synchronization has already revealed
a candidate node from which to fetch data.