Return DHASH_STALE for noauth writes that don't modify data.
Like keyhash returns STALE when you try to write a lower version number,
noauth should return STALE to indicate that the write didn't do anything
useful.

This is helpful for noticing if a replicated object's repair has "gone
the wrong way"---if the sync protocol doesn't know or can't tell which
different aux key is newer, it may queue a repair to send a stale
object to node with a newer object.  With this error response, the
repairing node can make sure that the node with the stale object gets
a copy of the newer one.