Switch replicated objects to use maintd for repair generation.
Add #ifdef'd code (enabled) that asks maintd for repairs.

Take advantage of source field now provided by maintd and DHash's
new DHASHCLIENT_SKIP_LOOKUP; encode the source location, if provided,
into rjrep's (and update the desc string appropriately).  Update
rjrep generation code for the old-sourceless case in noauth and keyhash
servers.

Add some handling for cases where the nodes involved in the repair
have stale data; DHASH_STALE is returned when to a store when the data
being written either the same as or older than what is on the node.
This is necessary because maintd does not try to aggregate merkle keys
by dbkey.  So, we no longer farm out work to real_repair in
subclasses.  We get a single repair for each merkle key and should
eventually converge to correctness based on the stale responses.

In the case where we detect that the source node's data is stale
relative to the source, enqueue a new repair that will try to push the
local copy out to the source node.  Similarly, if there is stale data
relative the destination, we try and read the newer version from the
destination and propogate it to the source.  Include a blocking
mechanism to check whether a repair is already in response to such a
situation; this is to avoid an infinite loop in the event that, for
example, both nodes have the same (up to date) data but stale sync
data causes one to begin a repair erroneously.

In general, it is not possible to tell which of two given
dbkey+auxkey's is newer so this approach is reasonable.  We might be
able to improve the server side response to distinguish between "the
same" and "old" and return a more nuanced error code (instead of just
DHASH_STALE to both cases).

Tweak logging to use modlogger instead of warn.