Align maintenance key ranges with dbkey ranges for mutable objects.
Since we mask out the low 32-bits of a chordID to be used as the
dbkey, we need to ensure that the rare (?) case of a node id
overlapping a dbkey range doesn't mess anything up.  This change
assigns the object to the successor of the masked out id and ensures
that all maintenance operations are divided between nodes along key
boundaries.  (The sync starts from and goes up to the clockwise side
of the 32-bit space.) This is particularly important so that a single
node learns about all the different variants of a mutable key.