Look-up mutable data by dbkey.
Export a new id_to_dbkey method in the block objects (in addition to
the one on the dhblock_replicated_srv class).  Use this to convert the
key to its dbkey before finding the successor list.  Looking up this
way helps establish consistency with the previous change to align key
sync ranges, hopefully avoiding the corner case where the lookup may
miss the "true" successor by one.   This change is server side and
transparent to the client...  mostly.

It is a little odd to be exposing the dbkey in the client interface
at all, even if it is only to the gateway side (dhashcli not
dhashclient).

Also, there are definitely issues of collisions in the dbkey space
(since we mask off 32 bits); confusing things will happen if there are
such collisions.  Technically there's no reason that the db key (used
for adbd) needs to be exactly 160 bits.  That restriction may hold for
merkley keys which is why we introduced it in the first place.  Still,
for now, be careful and cautious using NOAUTH (and KEYHASH) blocks.