Change vnode_impl::ping to call CHORDPROC_NULL.
The old ping code did not deal correctly with the case of a
physical node losing some vnodes; because it pinged using
TRANSPORTPROC_NULL, it would assume that the vnode existed,
mark it alive, only to discover later that there was no
corresponding vnode object on the other side and mark it
dead again.  Repeatedly.  This changes ping to make a call
at the CHORDPROC level, explicitly asking for the vnode
in question and treating the node as dead if the transport prot
returns an error.

Refactor marshalling code out of vnode_impl::doRPC so it can be
reused in the ping procedure.