Rewrite pmaint as a maintd global maintenance thread.
Global maintenance is designed to ensure that replicas and
fragments that wind up out of place for whatever reason have
an opportunity to return home, to where they can be found by
lookup.

The new global maintenance scheme works on batches of keys,
using Merkle sync in place of old DHASHPROC_OFFER.  Algorithm
works by incrementally scanning the local node's Merkle tree to
identify objects that seem out of place, find the current successor
for that key and then sync with that node to find keys that
should be sent to the remote node.

Technical details:
* Encapsulate global maint in maint_global object (maintainer friend).
* Enqueue up to count/2 global repairs as part of regular getrepairs.
* Tweak dhblock_*_srv::maintqueue + repair_job to handle global
  repairs differently; distinguish by whether key is in the replica range.

Still to consider:
* Delete objects after successful move?
* Delete objects that are sufficiently replicated?
* Improve efficiency of Merkle sync for this usage? (Half sync.)