Don't process repairs if too few predecessors.
This probably means that the node has been partitioned from all other
nodes in the system.   Failure to check this case was causing a SEGV
but it is a good idea anyway.  One day, maybe maintd will notice that
all neighbors are on the same host and not even generate the repairs
to begin with.